# docs/apifox/auth-api.openapi.yaml
openapi: 3.0.3
info:
  title: 智能抄表系统后端接口
  description: 当前认证、用户管理、表具管理、居民表具绑定、设备、告警、抄表任务及阿里云 OSS 抄表图片模块的 Apifox 测试文档。
  version: 2.8.0
servers:
  - url: http://localhost:8070
    description: 本地开发环境
...
paths:
...
  /api/v1/device/meter-reading-tasks/{taskId}/images:
    post:
      tags: [抄表图片]
      summary: 设备上传任务采集图片
      description: 设备通过 X-Device-No 和 X-Device-Secret 认证，只能给分配到本机的 PROCESSING 任务上传图片。
      operationId: uploadDeviceMeterImage
      security:
        - deviceNoHeader: []
          deviceSecretHeader: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                imageType:
                  $ref: '#/components/schemas/MeterImageType'
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传成功或命中幂等记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageItemSuccessResult'
        '401':
          description: 设备编号或密钥错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '409':
          description: 任务状态变化、达到图片上限或幂等标识冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '502':
          description: OSS 未配置或 OSS 操作失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
  /api/v1/admin/meter-images:
    get:
      tags: [抄表图片]
      summary: 管理员分页查询抄表图片
      operationId: listAdminMeterImages
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: {type: integer, minimum: 1, default: 1}
        - name: pageSize
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - name: taskNo
          in: query
          schema: {type: string, maxLength: 64}
        - name: meterNo
          in: query
          schema: {type: string, maxLength: 64}
        - name: uploaderType
          in: query
          schema: {$ref: '#/components/schemas/TaskExecutorType'}
        - name: uploaderId
          in: query
          schema: {type: integer, format: int64, minimum: 1}
        - name: imageStatus
          in: query
          schema: {$ref: '#/components/schemas/MeterImageStatus'}
        - name: storageStatus
          in: query
          schema: {$ref: '#/components/schemas/MeterImageStorageStatus'}
        - name: bound
          in: query
          schema: {type: boolean}
        - name: deleted
          in: query
          schema: {type: boolean}
        - name: createdAtStart
          in: query
          schema: {type: string, format: date-time}
        - name: createdAtEnd
          in: query
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImagePageSuccessResult'
  /api/v1/admin/meter-images/{imageId}:
    get:
      tags: [抄表图片]
      summary: 管理员查询图片详情
      operationId: getAdminMeterImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdPath'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageDetailSuccessResult'
        '404':
          description: 图片不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
  /api/v1/admin/meter-reading-tasks/{taskId}/images:
    get:
      tags: [抄表图片]
      summary: 管理员查询指定任务全部图片
      operationId: listAdminTaskMeterImages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: 查询成功，包含已软删除图片用于审计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageDetailListSuccessResult'
  /api/v1/admin/meter-images/{imageId}/invalidate:
    put:
      tags: [抄表图片]
      summary: 管理员标记图片无效
      operationId: invalidateMeterImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeterImageStatusDTO'
            example: {version: 0, reason: 图片模糊，无法审核}
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageVersionSuccessResult'
        '409':
          description: version 已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
  /api/v1/admin/meter-images/{imageId}/restore:
    put:
      tags: [抄表图片]
      summary: 管理员恢复图片
      operationId: restoreMeterImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeterImageStatusDTO'
            example: {version: 1, reason: 复核后确认图片有效}
      responses:
        '200':
          description: 恢复成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageVersionSuccessResult'
        '400':
          description: OSS 对象已删除或当前状态不能恢复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '409':
          description: version 已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceNoHeader:
      type: apiKey
      in: header
      name: X-Device-No
      description: 管理员创建设备时指定的全局唯一设备编号
    deviceSecretHeader:
      type: apiKey
      in: header
      name: X-Device-Secret
      description: 创建设备或重置密钥时仅返回一次的设备明文密钥
  parameters:
...
    TaskIdPath:
      name: taskId
      in: path
      required: true
      description: 抄表任务 ID
      schema:
        type: integer
        format: int64
        minimum: 1
    ImageIdPath:
      name: imageId
      in: path
      required: true
      description: 抄表图片元数据 ID
      schema:
        type: integer
        format: int64
        minimum: 1
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      description: 每次新上传生成一个唯一值；同一请求重试时保持不变，最大64个字符
      schema:
        type: string
        minLength: 1
        maxLength: 64
      example: 550e8400-e29b-41d4-a716-446655440000
  schemas:
...
    TaskExecutorType:
      type: string
      enum:
        - METER_READER
        - DEVICE
      description: 任务执行者类型，METER_READER 表示人工抄表员，DEVICE 表示采集设备
      example: METER_READER
...
    MeterImageType:
      type: string
      enum: [ORIGINAL, AI_ANNOTATED, ENVIRONMENT]
      default: ORIGINAL
      description: ORIGINAL 原始表具图片；AI_ANNOTATED AI 标注图片；ENVIRONMENT 现场环境图片
    MeterImageStatus:
      type: string
      enum: [VALID, INVALID]
      description: 图片业务有效性状态
    MeterImageStorageStatus:
      type: string
      enum: [STORED, DELETE_PENDING, DELETED]
      description: OSS 对象存储生命周期状态
    UpdateMeterImageStatusDTO:
      type: object
      required: [version, reason]
      properties:
        version:
          type: integer
          minimum: 0
          description: 当前图片乐观锁版本
        reason:
          type: string
          minLength: 1
          maxLength: 500
          description: 失效或恢复原因
    MeterImageItemVO:
      type: object
      required:
        - imageId
        - taskId
        - imageType
        - originalName
        - contentType
        - fileSize
        - imageStatus
        - storageStatus
        - version
        - createdAt
      properties:
        imageId: {type: integer, format: int64}
        taskId: {type: integer, format: int64}
        imageType: {$ref: '#/components/schemas/MeterImageType'}
        imageTypeName: {type: string}
        originalName: {type: string}
        contentType: {type: string, enum: [image/jpeg, image/png, image/webp]}
        fileSize: {type: integer, format: int64, description: 字节数}
        imageWidth: {type: integer, nullable: true}
        imageHeight: {type: integer, nullable: true}
        imageStatus: {$ref: '#/components/schemas/MeterImageStatus'}
        imageStatusName: {type: string}
        storageStatus: {$ref: '#/components/schemas/MeterImageStorageStatus'}
        resultId: {type: integer, format: int64, nullable: true}
        version: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
    MeterImageDetailVO:
      allOf:
        - $ref: '#/components/schemas/MeterImageItemVO'
        - type: object
          properties:
            taskNo: {type: string}
            taskStatus: {type: string}
            meterId: {type: integer, format: int64}
            meterNo: {type: string}
            meterName: {type: string}
            uploaderType: {$ref: '#/components/schemas/TaskExecutorType'}
            uploaderTypeName: {type: string}
            uploaderId: {type: integer, format: int64}
            uploaderCode: {type: string}
            uploaderName: {type: string}
            bucketName: {type: string}
            objectKey: {type: string}
            ossEtag: {type: string, nullable: true}
            sha256: {type: string}
            storageStatusName: {type: string}
            statusReason: {type: string, nullable: true}
            statusChangedBy: {type: integer, format: int64, nullable: true}
            statusChangedAt: {type: string, format: date-time, nullable: true}
            deleted: {type: boolean}
            deletedBy: {type: integer, format: int64, nullable: true}
            deletedAt: {type: string, format: date-time, nullable: true}
            updatedAt: {type: string, format: date-time}
    MeterImageVersionVO:
      type: object
      required: [imageId, imageStatus, version]
      properties:
        imageId: {type: integer, format: int64}
        imageStatus: {$ref: '#/components/schemas/MeterImageStatus'}
        version: {type: integer, minimum: 0}
...
    MeterImageItemSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data: {$ref: '#/components/schemas/MeterImageItemVO'}
        traceId: {type: string}
...
    MeterImageDetailSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data: {$ref: '#/components/schemas/MeterImageDetailVO'}
        traceId: {type: string}
    MeterImageDetailListSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data:
          type: array
          items: {$ref: '#/components/schemas/MeterImageDetailVO'}
        traceId: {type: string}
    MeterImagePageSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data:
          type: object
          required: [records, total, page, pageSize]
          properties:
            records:
              type: array
              items: {$ref: '#/components/schemas/MeterImageDetailVO'}
            total: {type: integer, format: int64}
            page: {type: integer}
            pageSize: {type: integer}
        traceId: {type: string}
    MeterImageVersionSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data: {$ref: '#/components/schemas/MeterImageVersionVO'}
        traceId: {type: string}
...
    ErrorResult:
      type: object
      required:
        - code
        - message
        - data
        - traceId
      properties:
        code:
          type: string
          example: AUTH_FAILED
        message:
          type: string
          example: 用户名或密码错误
        data:
          type: object
          nullable: true
          description: 错误响应时为 null
        traceId:
          type: string